+++ src/app/api/v1/customer/[id]/route.ts
@@ -1,5 +1,6 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
+import authService from "@/services/authService";
 
 let modelName = "Customer"
 export async function GET(
@@ -7,6 +8,16 @@ export async function GET(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     const data = await prisma.customer.findUnique({
       where: {
@@ -46,6 +57,16 @@ export async function PATCH(
   { params }: { params: { id: string }}
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     let json = await request.json();
 
@@ -57,7 +78,11 @@ export async function PATCH(
       await prisma.notification.create({
         data: {receiverId: updatedData.employeeId, resourceUrl: `/customers/${updatedData.id}`, message: `${updatedData.companyName} has been approved as your customer`}
       })
-    }
+    }/* else{
+      await prisma.notification.create({
+        data: {receiverId: updatedData.employeeId, resourceUrl: `/customers/${updatedData.id}`, message: `${updatedData.companyName} Customer details was edited`}
+      })
+    } */
 
 
     if (!updatedData) {
@@ -84,6 +109,16 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     await prisma.customer.delete({
       where: { id },
