+++ src/app/api/v1/marketingActivity/[id]/route.ts
@@ -1,6 +1,7 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import { uploadImage } from "@/services/imageService";
+import authService from "@/services/authService";
 
 let modelName = "Marketing Activity"
 export async function GET(
@@ -8,6 +9,16 @@ export async function GET(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     const data = await prisma.markettingActivity.findUnique({
       where: {
@@ -42,6 +53,16 @@ export async function PATCH(
   { params }: { params: { id: string }}
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     let json = await request.json();
     if(json?.newImages?.length > 0){
@@ -91,6 +112,16 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     await prisma.markettingActivity.delete({
       where: { id },
