+++ src/app/api/v1/pfiRequestForm/[id]/route.ts
@@ -1,5 +1,6 @@
 import { prisma } from "@/lib/prisma";
 import { NextResponse } from "next/server";
+import authService from "@/services/authService";
 
 let modelName = "Pfi Request Form"
 export async function GET(
@@ -7,6 +8,16 @@ export async function GET(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     const data = await prisma.pfiRequestForm.findUnique({
       where: {
@@ -45,27 +56,41 @@ export async function PATCH(
   { params }: { params: { id: string }}
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["admin", "supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     let json = await request.json();
+
+    const prevData = await prisma.pfiRequestForm.findUnique({
+      where: {id}
+    });
 
     const updatedData = await prisma.pfiRequestForm.update({
       where: { id },
       data: json,
     });
 
-    if(json.approved){
+    if(!prevData?.approved && json?.approved){
       await prisma.notification.create({
         data: { title: "Pfi Request", receiverId: updatedData.employeeId, resourceUrl: `/pfiRequests/${id}`, message: `Pfi with reference number: ${updatedData.pfiReferenceNumber} has been approved`}
       })
     }
 
-    if(json.locked){
+    if(!prevData?.locked && json?.locked){
       await prisma.notification.create({
         data: { title: "Pfi Request", staffCadre: "admin", resourceUrl: `/pfiRequests/${id}`, message: `New Pfi Request has been sent for approval`}
       })
     }
 
-    if(!json.locked){
+    if(!json?.locked && prevData?.locked){
       let pfiDetails = await prisma.pfiRequestForm.findUnique({where: {id}});
       await prisma.notification.create({
         data: { title: "Pfi Request", receiverId: pfiDetails?.employeeId, resourceUrl: `/pfiRequests/${id}`, message: `Pfi Request has been unlocked by admin`}
@@ -96,6 +121,16 @@ export async function DELETE(
   { params }: { params: { id: string } }
 ) {
   try {
+    const token = (request.headers.get("Authorization") || "").split("Bearer ").at(1) as string;
+    let {isAuthorized} = authService(token, ["supervisor", "salesPerson"])
+    if(!isAuthorized){
+      return new NextResponse(JSON.stringify({ message: `UnAuthorized`, data: null}), {
+        status: 401,
+        headers: { "Content-Type": "application/json" },
+      }); 
+    }
+
+
     const id = params.id;
     await prisma.pfiRequestForm.delete({
       where: { id },
